requirements.txt
alembic==1.16.4
amqp==5.3.1
async-timeout==5.0.1
asyncio==4.0.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
prompt_toolkit==3.0.51
protobuf==6.33.0
psycogreen==1.0.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
redis==6.4.0
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
upstox-python-sdk==2.17.0
urllib3==2.5.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
